+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 [13.8.12]
 - Fixed key buttons on Controller IO blocks
 - Fixed pick block not selecting matching hotbar framed blocks
-- Fixed particle icon of framed drawers not using material
+- NEO/FORGE: Fixed particle icon of framed drawers not using material
 
 [13.8.11]
 - Fixed rendering of framed drawer item blocks